enhance: Add put method to user endpoint

+++ MySqlDriver/src/main/java/org/example/mysql/database/UserMysqlRepository.java
@@ -3,5 +3,5 @@ package org.example.mysql.database;
 import org.example.mysql.schema.UserMysql;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserMysqlRepository extends CrudRepository<UserMysql, Integer> {
+public interface UserMysqlRepository extends CrudRepository<UserMysql, Long> {
 }
